==> ultron/1.2 trading_calendar_function.ipynb <==
获取交易日历，本例子使用沪深交易时间,还包括有商品期货和港股交易时间

from ultron.tradingday import *

begin_date ='2020-01-01'
end_date = '2022-03-10'

### 交易时间区间

##### 连续交易日

dates = makeSchedule(begin_date,end_date, '1b', 'china.sse',
                         BizDayConventions.Preceding)
dates

##### 每隔四天交易日

dates = makeSchedule(begin_date,end_date, '4b', 'china.sse',
                         BizDayConventions.Preceding)
dates

### 交易日推算

##### 向后推算最近的两个交易日

advanceDateByCalendar('china.sse', begin_date, '2b')

##### 向前推算最近的三个交易日

advanceDateByCalendar('china.sse', begin_date, '-3b')

### 判断是否为交易日

isBizDay('china.sse', begin_date)
