==> ultron/2.1 derived_factor_generation_signal.ipynb <==
在沙盒环境下，计算期货品种衍生因子信号

import os,pdb,itertools
os.environ['ULTRON_DATA'] = 'keim'

import pandas as pd
import numpy as np
from ultron.env import *
from ultron.tradingday import *
from ultron.strategy.experimental.single_factor import SingleFactor
from ultron.kdutils.parallel import delayed, Parallel

enable_example_env()

#### 选择中因子

sel_factor = pd.read_csv(os.path.join(g_project_data, 'sel_factor.csv'), index_col=0)
sel_factor.head()

#### 读取因子

total_data = pd.read_csv(os.path.join(g_project_data, 'factor.csv'), index_col=0)
factor_data = total_data[['trade_date','code'] + sel_factor['factor'].unique().tolist()]
factor_data.head()

#### 加载行情

market_data = pd.read_csv(os.path.join(g_project_data, 'market_data.csv'), index_col=0)
market_data.head()

#### 衍生因子计算

def _build(total_data, market_data, window, weekday, bins, columns):
    sf = SingleFactor(factor_data=None, market_data=None, codes=None, columns=None)
    normalize_data = sf.normalize(factor_data=total_data.dropna(subset=columns), 
                                    windows=window, columns=columns)
    normalize_data = normalize_data.sort_values(by=['trade_date','code'])
    normalize_data['trade_date'] = pd.to_datetime(normalize_data['trade_date'])
    dates = makeSchedule(normalize_data.trade_date.min(), 
                    normalize_data.trade_date.max(),
                    '1b', 'china.sse', BizDayConventions.Preceding)
    dates = [d for d in dates if d.weekday() == (weekday-1)]
    normalize_data = normalize_data.set_index('trade_date').loc[dates].reset_index()
    returns_data = sf.returns(market_data.set_index(['trade_date','code']), period=1)
    period_data = sf._transformer(normalize_data=normalize_data, 
                                    returns=returns_data, columns=columns, 
                                    period='1b')
    period_data['trade_date'] = pd.to_datetime(period_data['trade_date'])
    returns_data['trade_date'] = pd.to_datetime(returns_data['trade_date'])
    total_data = returns_data.merge(period_data, on=['trade_date','code'])

    #res = {}
    res = []
    for column in columns:
        position = sf.quantile(normalize_data=period_data, factor_name=column, n_bins=bins).reset_index()
        position['signal'] = np.where(position['group']==bins, 1, np.where(position['group']==1,
                                               -1,0))
        dct = {'position':position,
                'params':{'window':window,'weekday':weekday,'bins':bins},
                'factor':column}
        res.append(dct)
    return res

def transfer(res):
    new_name = res['factor'] + '_' + str(res['params']['window']) + '_' + str(
        res['params']['weekday']) + '_' + str(res['params']['bins'])
    rt = res['position'][['trade_date','code','signal']].rename(columns={'signal':new_name})
    return rt.set_index(['trade_date','code'])

parallel = Parallel(n_jobs=4, verbose=1, pre_dispatch='2*n_jobs')

#### 计算衍生因子

out = parallel(delayed(_build)(factor_data[['trade_date','code', v['factor']]], market_data,  
                   window=v['window'], weekday=v['weekday'], 
                   bins=v['bins'], columns=[v['factor']]) for v in sel_factor.to_dict(orient='records'))

results = list(itertools.chain.from_iterable(out))

res = [transfer(r) for r in results]

#### 生成信号数据

signal_dt = pd.concat(res, axis=1).reset_index()
signal_dt['trade_date'] = pd.to_datetime(signal_dt['trade_date'])
signal_dt = signal_dt.sort_values(by=['trade_date','code'])
signal_dt.tail()

#### 保存信号数据 用于其他例子使用

signal_dt.to_csv(os.path.join(g_project_data, 'signal.csv'),encoding='UTF-8')
