==> ultron/2.2 dimension_reduction_multidimension_features.ipynb <==
在沙盒环境，对多特征进行降维的方法

import sys, os
import pandas as pd
import numpy as np
os.environ['ULTRON_DATA'] = 'minden'
os.environ['IGNORE_WARNINGS'] = '0'

sys.path.insert(0, os.path.abspath('../'))

from ultron.env import *
from ultron.factor.dimension import DimensionLDA
from ultron.factor.dimension import DimensionPCA
from ultron.factor.dimension import DimensionKeans
from ultron.factor.dimension import DimensionCorrcoef

indicator_data = pd.read_csv(os.path.join(g_project_data, 'indicator.csv'), index_col=0)
indicator_data['trade_date'] = pd.to_datetime(indicator_data['trade_date'])
indicator_data.head()

from sklearn.datasets import make_classification
X, y = make_classification(n_samples=1000, n_features=20, n_informative=15, n_redundant=5, 
                           random_state=7, n_classes=10)
samples_data = pd.DataFrame(X)
samples_data.head()

indicator_features = [col for col in indicator_data.columns if col not in ['code','trade_date','key','date_week']]
samples_features = [col for col in samples_data.columns]

采用KEANS对数据进行聚类，通过聚类的结果对数据进行降维，其中对聚类结果提供三种校验方式
- 1. squares
- 2. calinski_harabas
- 3. silhoette

#### LDA 降维

indicator 数据 此处采用squares评估

kmeans = DimensionKeans(method='squares', features=indicator_features, thresh=0.9)
data = kmeans.run(factors_data=indicator_data)
lda = DimensionLDA(features=indicator_features)
lda.run(data, 9)

sample 数据 此处采用harabasz评估

kmeans = DimensionKeans(method='harabasz', features=samples_features, thresh=0.9)
data = kmeans.run(factors_data=samples_data)
lda = DimensionLDA(features=samples_features)
lda.run(data, 19)

#### PCA 降维

samples 数据 此处采用squares评估

kmeans = DimensionKeans(method='squares', features=samples_features, thresh=0.9)
data = kmeans.run(factors_data=samples_data)
pca = DimensionPCA(features=samples_features)
pca.run(data, 9)

#### 相关性降维

from ultron.factor.dimension.corrcoef import FCorrType
from ultron.ump.similar.corrcoef import ECoreCorrType

##### 横截面降维

#### 皮尔逊
engine = DimensionCorrcoef(features=indicator_features, thresh=0.97, method=FCorrType.F_CS_CORR)
engine.run(factors_data=indicator_data,similar_type=ECoreCorrType.E_CORE_TYPE_PEARS)

#### 斯皮尔曼
engine = DimensionCorrcoef(features=indicator_features, thresh=0.97, method=FCorrType.F_CS_CORR)
engine.run(factors_data=indicator_data,similar_type=ECoreCorrType.E_CORE_TYPE_SPERM)

#### 基于PEARS使用序列＋－符号
engine = DimensionCorrcoef(features=indicator_features, thresh=0.97, method=FCorrType.F_CS_CORR)
engine.run(factors_data=indicator_data,similar_type=ECoreCorrType.E_CORE_TYPE_SIGN)

#### 基于PEARS使用移动时间加权
engine = DimensionCorrcoef(features=indicator_features, thresh=0.85, method=FCorrType.F_CS_CORR)
engine.run(factors_data=indicator_data,similar_type=ECoreCorrType.E_CORE_TYPE_ROLLING, window=20)

##### 时序降维

#### 皮尔逊
engine = DimensionCorrcoef(features=indicator_features, thresh=0.97, method=FCorrType.F_TS_CORR)
engine.run(factors_data=indicator_data,similar_type=ECoreCorrType.E_CORE_TYPE_PEARS)

#### 斯皮尔曼
engine = DimensionCorrcoef(features=indicator_features, thresh=0.97, method=FCorrType.F_TS_CORR)
engine.run(factors_data=indicator_data,similar_type=ECoreCorrType.E_CORE_TYPE_SPERM)

#### 基于PEARS使用序列＋－符号
engine = DimensionCorrcoef(features=indicator_features, thresh=0.97, method=FCorrType.F_TS_CORR)
engine.run(factors_data=indicator_data,similar_type=ECoreCorrType.E_CORE_TYPE_SIGN)
