==> ultron/3.3 futures_features_machine_learning_optimizer.ipynb <==
在沙盒环境，多特征非线性合成及进行组合优化

import os,pdb,itertools,copy,datetime
os.environ['ULTRON_DATA'] = 'keim'

import pandas as pd
import numpy as np
from ultron.factor.data.processing import factor_processing
from ultron.factor.data.standardize import standardize
from ultron.optimize.model.treemodel import RandomForestRegressor
from ultron.optimize.model.linearmodel import LinearRegression
from ultron.strategy.experimental.single_factor import SingleFactor
from ultron.strategy.fut_strategy import create_params
from ultron.strategy.strategy import StrategyEngine
from ultron.kdutils.parallel import delayed, Parallel

from ultron.env import *

enable_example_env()

#### 加载行情数据

market_data = pd.read_csv(os.path.join(g_project_data, 'market_data.csv'), index_col=0)
market_data['trade_date'] = pd.to_datetime(market_data['trade_date'])
market_data.head()

#### 选择中因子

sel_factor = pd.read_csv(os.path.join(g_project_data, 'sel_factor.csv'), index_col=0)
sel_factor = sel_factor.drop_duplicates(subset=['factor','window'])
sel_factor = sel_factor[(sel_factor['factor'] != 'inventory')]
sel_factor = sel_factor[(sel_factor['factor'] != 'profitratio')]
sel_factor.head()

#### 读取因子

total_data = pd.read_csv(os.path.join(g_project_data, 'factor.csv'), index_col=0)
total_data['trade_date'] = pd.to_datetime(total_data['trade_date'] )
factor_data = total_data[['trade_date','code'] + sel_factor['factor'].unique().tolist()]
factor_data.head()

#### 加载行业信息

industry_data = pd.read_csv(os.path.join(g_project_data, 'industry_data.csv'), index_col=0)
industry_data.head()

#### 时序标准化

parallel = Parallel(n_jobs=4, verbose=1, pre_dispatch='2*n_jobs')

def _build(total_data, window, columns):
    sf = SingleFactor(factor_data=None, market_data=None, codes=None, columns=None)
    normalize_data = sf.normalize(factor_data=total_data.dropna(subset=columns), 
                                    windows=window, columns=columns)
    normalize_data = normalize_data.sort_values(by=['trade_date','code'])
    return normalize_data.set_index(['trade_date','code']).rename(
        columns={columns[0]:columns[0] + '_' + str(window)})

out = parallel(delayed(_build)(factor_data[['trade_date','code', v['factor']]],
                   window=v['window'], columns=[v['factor']]) for v in sel_factor.to_dict(orient='records'))

factors_data = pd.concat(out, axis=1).reset_index().fillna(0)
factors_data.head()

#### 横截面标准化

期货品种不做去极值和中性化

diff_cols = ['trade_date','code']
features = [col for col in factors_data.columns if col not in diff_cols]

alpha_res = []
grouped = factors_data.groupby(['trade_date'])
for k, g in grouped:
    new_factors = factor_processing(g[features].values,
                 pre_process=[standardize])
    f = pd.DataFrame(new_factors, columns=features)
    for k in diff_cols:
        f[k] = g[k].values
    alpha_res.append(f)

factors_data = pd.concat(alpha_res)
factors_data.tail()

构建模型X,Y值

##### 目前使用收益率为第二天的收益率为Y，来构建模型

def returns(market_data):
    price_tb = market_data['closePrice'].unstack()
    price_tb.fillna(method='pad', inplace=True)
    return_tb = np.log(price_tb.shift(-1) / price_tb)
    return_tb = return_tb.replace([np.inf, -np.inf], np.nan)
    return_tb = return_tb.stack().reindex(market_data.index)
    return_tb.name = 'nxt1_ret'
    return return_tb

rets = returns(market_data.set_index(['trade_date','code']))
rets.head()

train_data = factors_data.merge(rets, on=['trade_date','code'])
train_data = train_data.dropna(subset=['nxt1_ret']).fillna(0)
train_data.head()

可使用多个机器学习模型，这里为了执行效率，使用线性回归模型

#model = RandomForestRegressor(features=features)
model = LinearRegression(features=features, fit_intercept=False)

%%time
model.fit(train_data, train_data['nxt1_ret'].values)

#### 组合构建

此处暂时使用等权为初始权重

def _weighted(data, equal=1):
    data['factor'] =1
    weighted =  1 / len(data)
    weighted = pd.DataFrame([weighted for i in range(0, len(data))],
                                columns=['factor'],
                                index=data.index)
    weighted['code'] = data['code'].values
    weighted = weighted.reset_index().drop(['trade_date'],axis=1)
    return weighted.rename(columns={'factor':'weight'})
benchmark_weighed = factors_data[['trade_date','code']].set_index('trade_date').groupby(
    level=['trade_date']).apply(
    lambda x: _weighted(x)).reset_index().drop(['level_1'],axis=1)
benchmark_weighed.head()

industry_dummy = pd.get_dummies(industry_data.set_index('code')['industry']).reset_index()
industry_dummy.head()

factors_data = factors_data.merge(
    industry_data, on=['code']).merge(
    industry_dummy, on=['code']).set_index(['trade_date','code'])
factors_data = factors_data.reset_index()
factors_data.tail()

neutralized_styles = factors_data['industry'].unique().tolist()

factors_data = factors_data.merge(
    benchmark_weighed, on=['trade_date','code']).sort_values(
    by=['trade_date','code'],ascending=True
)

factors_data = factors_data.rename(columns={'industry':'industry_code'})
is_in_benchmark = (factors_data.weight != 0.).astype(float).values.reshape((-1, 1))
factors_data.loc[:, 'benchmark'] = is_in_benchmark
factors_data.loc[:, 'total'] = np.ones_like(is_in_benchmark)
factors_data.tail()

factors_data = factors_data.merge(rets.reset_index(), on=['trade_date','code']).rename(
    columns={'nxt1_ret':'pev1_ret'})
factors_data.head()

effective_codes = factors_data.industry_code.unique().tolist() ### 有效行业
invalid_codes = [] ### 无效行业
target_vol = 0.05 ### 目标波动率
turn_over_target = 0.4  ### 换手率
lbound = -0.1 ### 单个标的优化下限
ubound = 0.1 ### 单个标的优化上限
benchmark_lower = -1.00 ## 基于基准下限
benchmark_upper = 1.001 ## 基于基准上限
total_lower = -0.01 ## 组合下限
total_upper = 0.01 ## 组合上限
effective_industry_lower = -0.1 ## 有效行业下限
effective_industry_upper = 0.1 ## 有效行业上限
invalid_industry_lower = -0.2 ## 无效行业下限
invalid_industry_lower = 0.2 ## 无效行业上限
cov_method = 'unshrunk' ## 协方差收益率模式
method = 'long_short'

params = create_params(
    effective_codes=effective_codes, invalid_codes=invalid_codes,
    target_vol=target_vol,turn_over_target=turn_over_target, 
    lbound=lbound, ubound=ubound,
    benchmark_lower=benchmark_lower,benchmark_upper=benchmark_upper,
    total_lower=total_lower,total_upper=total_upper,
    effective_industry_lower=effective_industry_lower,
    effective_industry_upper=effective_industry_upper,
    method=method,
    is_benchmark=1)

#### 构建模型

models = {}
factors_data_groups = factors_data.groupby('trade_date')
for ref_date, _ in factors_data_groups:
    models[ref_date] = model

strategy = StrategyEngine.create_class(name='fut')(alpha_model=models, total_data=factors_data, 
                    start_date=datetime.datetime(2021,1,4))

positions = strategy.create_positions(params=params)
