==> ultron/4.1  stock_ump_construction_features.ipynb <==
在沙盒环境，构建UMP自定义特征

import os, copy, sys
os.environ['ULTRON_DATA'] = 'minden'
sys.path.insert(0, os.path.abspath('../'))
import pandas as pd
import numpy as np

from ultron.ump.trade import ml_feature
from ultron.kdutils.progress import Progress
from ultron.kdutils.parallel import Parallel,delayed
from ultron.ump.trade.ml_feature import FeatureBase, BuyFeatureMixin, SellFeatureMixin
from ultron.ump.technical.vwap import calc_vwap

from ultron.env import *
from ultron.ump.core import env

enable_example_env()

env.g_enable_ml_feature = True

#### 加载指标数据

indicator_data = pd.read_csv(os.path.join(g_project_data, 'indicator.csv'), index_col=0)
indicator_data['trade_date'] = pd.to_datetime(indicator_data['trade_date'])
indicator_data = indicator_data.rename(columns={'preClosePrice':'pre_close','openPrice':'open',
                        'closePrice':'close','highestPrice':'high','lowestPrice':'low',
                        'turnoverVol':'volume'}).drop(['turnoverValue'],axis=1).set_index(['trade_date'])
indicator_data['date'] = indicator_data.index
indicator_data['p_change'] = indicator_data['p_change'] * 100
indicator_data.head()

#### 加载交易记录

trader_data = pd.read_csv(os.path.join(g_project_data, 'trader_records.csv'), index_col=0)
trader_data['buy_date'] = pd.to_datetime(trader_data['buy_date'])
trader_data['sell_date'] = pd.to_datetime(trader_data['sell_date'])
trader_data.head()

#### 自定义构建 VWAP特征

class FeatureVWap(FeatureBase, BuyFeatureMixin, SellFeatureMixin):
    """vwap特征，支持买入，卖出"""

    def __init__(self):
        self.vwap_keys = frozenset([21, 42, 60])

    def get_feature_keys(self, buy_feature):
        return [
            '{}vwap_ang{}'.format(self.feature_prefix(buy_feature=buy_feature),
                                  dk) for dk in self.vwap_keys
        ]

    def calc_feature(self, kl_pd, combine_kl_pd, day_ind, buy_feature):
        """
        根据买入或者卖出时的金融时间序列，以及交易日信息构造拟合波动率特征
        :param kl_pd: 择时阶段金融时间序列
        :param combine_kl_pd: 合并择时阶段之前1年的金融时间序列
        :param day_ind: 交易发生的时间索引，即对应self.kl_pd.key
        :param buy_feature: 是否是买入特征构造（bool）
        :return: 构造波动率特征的键值对字典
        """
        vwap_dict = {}
        for dk in self.vwap_keys:
            # 迭代预设角度周期，计算构建特征
            if day_ind - dk >= 0:
                # 如果择时时间序列够提取特征，使用kl_pd截取特征交易周期收盘价格
                vwap_close = kl_pd[day_ind - dk + 1:day_ind + 1]  #.close
            else:
                # 如果择时时间序列不够提取特征，使用combine_kl_pd截取特征交易周期，首先截取直到day_ind的时间序列
                combine_kl_pd = combine_kl_pd.loc[:kl_pd.index[day_ind]]
                # 如combine_kl_pd长度大于特征周期长度－> 截取combine_kl_pd[-dk:].close，
                # 否则取combine_kl_pd所有交易收盘价格
                vwap_close = combine_kl_pd[
                    -dk:] if combine_kl_pd.shape[0] > dk else combine_kl_pd
            ##
            vwap_price = calc_vwap(vwap_close)
            vwap_score = vwap_price.score
            # 标准vwap价格特征值
            vwap_score = 0 if np.isnan(vwap_score) else round(vwap_score, 3)
            # vwap特征键值对字典添加波动特征key和对应的波动特征值
            vwap_dict['{}vwap_ang{}'.format(
                self.feature_prefix(buy_feature=buy_feature), dk)] = vwap_score
        return vwap_dict

#### 将自定义特征加入计算队列

ml_feature.clear_user_feature()
ml_feature.append_user_feature(FeatureVWap)

#### 创建特征计算对象

mf = ml_feature.MlFeature()

#### 每支股票计算单方向特征

import pdb
def calc_features(indicator_data, code, dates, buy_feature):
    alpha_res = []
    keys = indicator_data.loc[dates]['key'].values
    ## 暂时过滤0
    keys = [key - 1 for key in keys if key!=0]
    indicator_data = copy.deepcopy(indicator_data)
    if len(keys) < 1:
        return pd.DataFrame()
    for key in keys:
        rt = mf.make_feature_dict(kl_pd=indicator_data, combine_kl_pd=indicator_data, day_ind=key, 
                                  buy_feature=buy_feature)
        rt['key'] = key + 1 ## 此处为了模型训练 + 1
        alpha_res.append(rt)
        result = pd.DataFrame(alpha_res)
    result = result.merge(indicator_data[['key','date','code']], on=['key'])
    return result.drop(['key'],axis=1)

#### 创建每支股票开平特征

def create_features(code, indicator_data):
    trader_dt = trader_data.set_index('code').loc[[code]]
    if len(trader_dt) < 1:
        return
    buy_dates = trader_dt['buy_date'].dt.strftime('%Y-%m-%d').unique().tolist()
    sell_dates =  trader_dt['sell_date'].dt.strftime('%Y-%m-%d').unique().tolist()
    indicator_data = indicator_data.reset_index().set_index('code').loc[code].reset_index()
    indicator_data['key'] = indicator_data.index
    indicator_data = indicator_data.set_index('trade_date')
    buy_feature = calc_features(indicator_data, code, buy_dates, buy_feature=True)
    sell_feature = calc_features(indicator_data, code, sell_dates, buy_feature=False)
    name = "{0}.h".format(code)
    return {'name':code, 'sell': sell_feature, 'buy':buy_feature}
    #h5 = pd.HDFStore('../data/{0}/feature/'.format(task_id) + name,'w', complevel=4, complib='blosc')
    #h5['sell'] = sell_feature
    #h5['buy'] = buy_feature
    #h5.close()

i = 0
res = []
codes = trader_data['code'].unique().tolist()
with Progress(len(codes), 0, 'feature progress') as pg:
    for code in codes:
        i += 1
        pg.show(i,ext='{0}'.format(code))
        rt = create_features(code, indicator_data)
        res.append(rt)

buy_res = []
sell_res = []
for r in res:
    buy_res.append(r['buy'])
    sell_res.append(r['sell'])

buy_features = pd.concat(buy_res,axis=0)
sell_features = pd.concat(sell_res,axis=0)

##### 开仓特征

buy_features.head()

#### 平仓特征

sell_features.head()

#### 保存特征数据 用于其他例子使用

buy_features.to_csv(os.path.join(g_project_data, 'buy_features.csv'),encoding='UTF-8')
sell_features.to_csv(os.path.join(g_project_data, 'sell_features.csv'),encoding='UTF-8')
