==> ultron/4.4 predicted_custom_intercept.ipynb <==
在沙盒环境，构建UMP主裁模型和边裁模型混合模型

import os, copy
os.environ['ULTRON_DATA'] = 'minden'
import pandas as pd
import numpy as np
from sklearn import metrics

from ultron.env import *
os.environ['UMP_MODEL_PATH'] = g_project_data

from ultron.ump.core import env
from ultron.ump.ump import manager
from ultron.ump.ump.main_base import MainBase
from ultron.ump.ump.main_mul import MainMul
from ultron.ump.ump.base import BuyUmpMixin, ump_main_make_xy
from ultron.ump.model.principles import Principles
from ultron.ump.trade.ml_feature import FeatureBase, BuyFeatureMixin, SellFeatureMixin
from ultron.ump.technical.vwap import calc_vwap
from ultron.ump.trade import ml_feature

from ultron.ump.ump.edge_deg import EdgeDeg
from ultron.ump.ump.edeg_price import EdgePrice
from ultron.ump.ump.edge_wave import EdgeWave
from ultron.ump.ump.edge_full import EdgeFull

from ultron.ump.ump.main_full import MainFull
from ultron.ump.ump.main_price import MainPrice
from ultron.ump.ump.main_wave import MainWave
from ultron.ump.ump.main_deg import MainDeg

#### 加载买入特征

buy_features = pd.read_csv(os.path.join(g_project_data, 'buy_features.csv'), index_col=0)
buy_features['date'] = pd.to_datetime(buy_features['date'])
buy_features.head()

#### 加载卖出特征

sell_features = pd.read_csv(os.path.join(g_project_data, 'sell_features.csv'), index_col=0)
sell_features['date'] = pd.to_datetime(sell_features['date'])
sell_features.head()

#### 加载交易记录

trader_data = pd.read_csv(os.path.join(g_project_data, 'trader_records.csv'), index_col=0)
trader_data['buy_date'] = pd.to_datetime(trader_data['buy_date'])
trader_data['sell_date'] = pd.to_datetime(trader_data['sell_date'])
trader_data.head()

#### 合并数据集生成订单集

orders_pd = trader_data.merge(
    buy_features.rename(columns={'date': 'buy_date'}),
    on=['buy_date',
        'code']).merge(sell_features.rename(columns={'date': 'sell_date'}))
orders_pd.head()

diff_cols = ['code','buy_date','buy_cnt','sell_date','buy_price','sell_price']
cols = [col for col in orders_pd.columns if col not in diff_cols]

#### 创建数据集

orders_pd['profit'] = orders_pd['sell_price'] - orders_pd['buy_price']
orders_pd['profit_cg'] = (orders_pd['sell_price'] - orders_pd['buy_price']) / (
    orders_pd['buy_price'] * orders_pd['buy_cnt'])
orders_pd['profit_cg_hunder'] = orders_pd['profit_cg'] * 100
orders_pd['keep_days'] = (orders_pd['sell_date'] -
                          orders_pd['buy_date']).apply(lambda x: x.days)
orders_pd['result'] = np.where(orders_pd['profit'] > 0, 1, -1)
orders_pd['trade_date'] = orders_pd['buy_date']
orders_pd['key'] = orders_pd.index

orders_pd['ml_features'] = orders_pd.apply(lambda x: x[cols].to_dict(), axis=1)
orders_pd = orders_pd.set_index('trade_date')
orders_pd.head()

#### 自定义特征

class FeatureVWap(FeatureBase, BuyFeatureMixin, SellFeatureMixin):
    """vwap特征，支持买入，卖出"""

    def __init__(self):
        self.vwap_keys = frozenset([21, 42, 60])

    def get_feature_keys(self, buy_feature):
        return [
            '{}vwap_ang{}'.format(self.feature_prefix(buy_feature=buy_feature),
                                  dk) for dk in self.vwap_keys
        ]

    def calc_feature(self, kl_pd, combine_kl_pd, day_ind, buy_feature):
        """
        根据买入或者卖出时的金融时间序列，以及交易日信息构造拟合波动率特征
        :param kl_pd: 择时阶段金融时间序列
        :param combine_kl_pd: 合并择时阶段之前1年的金融时间序列
        :param day_ind: 交易发生的时间索引，即对应self.kl_pd.key
        :param buy_feature: 是否是买入特征构造（bool）
        :return: 构造波动率特征的键值对字典
        """
        vwap_dict = {}
        for dk in self.vwap_keys:
            # 迭代预设角度周期，计算构建特征
            if day_ind - dk >= 0:
                # 如果择时时间序列够提取特征，使用kl_pd截取特征交易周期收盘价格
                vwap_close = kl_pd[day_ind - dk + 1:day_ind + 1]  #.close
            else:
                # 如果择时时间序列不够提取特征，使用combine_kl_pd截取特征交易周期，首先截取直到day_ind的时间序列
                combine_kl_pd = combine_kl_pd.loc[:kl_pd.index[day_ind]]
                # 如combine_kl_pd长度大于特征周期长度－> 截取combine_kl_pd[-dk:].close，
                # 否则取combine_kl_pd所有交易收盘价格
                vwap_close = combine_kl_pd[
                    -dk:] if combine_kl_pd.shape[0] > dk else combine_kl_pd
            ##
            vwap_price = calc_vwap(vwap_close)
            vwap_score = vwap_price.score
            # 标准vwap价格特征值
            vwap_score = 0 if np.isnan(vwap_score) else round(vwap_score, 3)
            # vwap特征键值对字典添加波动特征key和对应的波动特征值
            vwap_dict['{}vwap_ang{}'.format(
                self.feature_prefix(buy_feature=buy_feature), dk)] = vwap_score
        return vwap_dict

#### 自定义主裁模型

class MainMul(MainBase, BuyUmpMixin):

    class MulFiter(Principles):

        @ump_main_make_xy
        def make_xy(self, **kwarg):
            regex = 'result|{}|{}|{}|{}|{}'.format(
                ml_feature.FeatureDeg().get_feature_ump_keys(
                    ump_cls=MainMul)[-1],
                ml_feature.FeaturePrice().get_feature_ump_keys(
                    ump_cls=MainMul)[-1],
                ml_feature.FeatureWave().get_feature_ump_keys(
                    ump_cls=MainMul)[-1],
                ml_feature.FeatureAtr().get_feature_ump_keys(
                    ump_cls=MainMul)[-1],
                FeatureVWap().get_feature_ump_keys(
                    ump_cls=MainMul)[-1],
            )
            mul_df = self.order_has_ret.filter(regex=regex)
            return mul_df

    def get_predict_col(self):
        return [
            ml_feature.FeatureDeg().get_feature_ump_keys(ump_cls=MainMul)[-1],
            ml_feature.FeaturePrice().get_feature_ump_keys(
                ump_cls=MainMul)[-1],
            ml_feature.FeatureWave().get_feature_ump_keys(ump_cls=MainMul)[-1],
            ml_feature.FeatureAtr().get_feature_ump_keys(ump_cls=MainMul)[-1],
            FeatureVWap().get_feature_ump_keys(ump_cls=MainMul)[-1]
        ]

    def get_fiter_class(self):
        return MainMul.MulFiter

    @classmethod
    def class_unique_id(cls):
        """
        具体ump类关键字唯一名称，类方法：return 'mul_main'
        主要针对外部user设置自定义ump使用, 需要user自己保证class_unique_id的唯一性，内部不做检测
        具体使用见ABuUmpManager中extend_ump_block方法
        """
        return 'mul_main'

#### 创建裁决模型

main_deg = MainDeg(predict=True, market_name='cn')
main_price = MainPrice(predict=True, market_name='cn')
main_wave = MainWave(predict=True, market_name='cn')
main_full = MainFull(predict=True, market_name='cn')
main_mul = MainMul(predict=True, market_name='cn')

edge_deg = EdgeDeg(predict=True, market_name='cn')
edge_price = EdgePrice(predict=True, market_name='cn')
edge_wave = EdgeWave(predict=True, market_name='cn')
edge_full = EdgeFull(predict=True, market_name='cn')

#### 结果判断

def apply_ml_features_edge(order, predicter):
    if not isinstance(order.ml_features, dict):
        # 低版本pandas dict对象取出来会成为str
        ml_features = ast.literal_eval(order.ml_features)
    else:
        ml_features = order.ml_features
    # 边裁进行裁决
    edge = predicter.predict(**ml_features)
    return edge.value

def apply_ml_features_main(order, predicter):
    if not isinstance(order.ml_features, dict):
        # 低版本pandas dict对象取出来会成为str
        ml_features = ast.literal_eval(order.ml_features)
    else:
        ml_features = order.ml_features
    # 主裁进行裁决
    edge = predicter.predict_kwargs(**ml_features)
    return -1 if edge == 1 else 0

orders_pd['main_deg'] = orders_pd.apply(apply_ml_features_main,
                                        axis=1,
                                        args=(main_deg, ))

orders_pd['main_full'] = orders_pd.apply(apply_ml_features_main,
                                         axis=1,
                                         args=(main_full, ))

orders_pd['main_wave'] = orders_pd.apply(apply_ml_features_main,
                                         axis=1,
                                         args=(main_wave, ))

orders_pd['main_price'] = orders_pd.apply(apply_ml_features_main,
                                          axis=1,
                                          args=(main_price, ))


orders_pd['main_mul'] = orders_pd.apply(apply_ml_features_main,
                                          axis=1,
                                          args=(main_mul, ))

orders_pd['edge_deg'] = orders_pd.apply(apply_ml_features_edge,
                                        axis=1,
                                        args=(edge_deg, ))

orders_pd['edge_price'] = orders_pd.apply(apply_ml_features_edge,
                                          axis=1,
                                          args=(edge_price, ))

orders_pd['edge_wave'] = orders_pd.apply(apply_ml_features_edge,
                                         axis=1,
                                         args=(edge_wave, ))

orders_pd['edge_full'] = orders_pd.apply(apply_ml_features_edge,
                                         axis=1,
                                         args=(edge_full, ))

### 统计成功率

#### 非均衡技术
def accuracy_score(name):
    accuracy = 0
    sub_edge_block_pd = orders_pd[(orders_pd[name] == -1)]
    if sub_edge_block_pd.shape[0] > 0:
        accuracy = metrics.accuracy_score(sub_edge_block_pd[name],
                                          sub_edge_block_pd.result)
    return accuracy

res = []
main_names = ['main_deg', 'main_price', 'main_wave', 'main_full', 'main_mul']
edge_names = ['edge_deg', 'edge_price', 'edge_wave', 'edge_full']
names = main_names + edge_names

for name in names:
    score = accuracy_score(name=name)
    res.append({'name': name, 'score': score})

weighted = pd.DataFrame(res)

### 权重换算
def transfor_weighted(weighted, names):
    sub_weighted = weighted.set_index('name').loc[names]
    return sub_weighted / sub_weighted.sum()

main_weighted = transfor_weighted(weighted, main_names) * 0.6 ### 人工赋主裁权重
edge_weighted = transfor_weighted(weighted, edge_names) * 0.4 ### 人工赋边裁权重
weighted = pd.concat([main_weighted, edge_weighted], axis=0)
weighted = dict(zip(weighted.index, weighted.score))

def apply_ump(order, weighted):
    result = 0
    for name in names:
        result += order[name] * weighted[name]
    return result

orders_pd['ump'] = orders_pd.apply(apply_ump, axis=1, args=(weighted, ))

orders_pd[['code','buy_date','buy_cnt','ump']]
