==> ultron/5.1  time_series_machine_learning.ipynb <==
在沙盒环境，使用机器学习构建时序涨跌幅预测模型

import os,pdb,itertools,copy,datetime
os.environ['ULTRON_DATA'] = 'keim'

import pandas as pd
import numpy as np
from ultron.env import *
from ultron.optimize.model.treemodel import XGBTrainer

enable_example_env()

#### 加载行情数据

market_data = pd.read_csv(os.path.join(g_project_data, 'market_data.csv'), index_col=0)
market_data['trade_date'] = pd.to_datetime(market_data['trade_date'])
market_data.head()

#### 选择中因子

sel_factor = pd.read_csv(os.path.join(g_project_data, 'sel_factor.csv'), index_col=0)
sel_factor.head()

#### 读取因子

total_data = pd.read_csv(os.path.join(g_project_data, 'factor.csv'), index_col=0)
factor_data = total_data[['trade_date','code'] + sel_factor['factor'].unique().tolist()]
factor_data['trade_date'] = pd.to_datetime(factor_data['trade_date'])
factor_data.head()

features = [col for col in factor_data.columns if col not in ['trade_date','code','inventory','profitratio']]

#### 构建训练集Y值

##### 通过ATR进行涨跌幅判断

def ATR(high, low, close, N=14):
    tmp1 = high - low
    tmp2 = (high - close.shift()).abs()
    tmp3 = (low - close.shift()).abs()
    cond1 = tmp1 > tmp2
    tmp2[cond1] = tmp1[cond1]
    cond2 = tmp2 > tmp3
    tmp3[cond2] = tmp2[cond2]
    TR = tmp3
    ATR = TR.rolling(N).mean()
    return ATR

def create_lable(market_data):
    market_data = market_data.sort_values(by=['trade_date', 'code'],
                                              ascending=True)
    highest_price = market_data.set_index(['trade_date', 'code'
                                               ])['highestPrice'].unstack()
    lowest_price = market_data.set_index(['trade_date', 'code'
                                              ])['lowestPrice'].unstack()
    close_price = market_data.set_index(['trade_date',
                                             'code'])['closePrice'].unstack()
    atr = ATR(high=highest_price,
                       low=lowest_price,
                       close=close_price,
                       N=20)
    signal_dt = (close_price - close_price.shift(1)) / atr
    signal_dt = signal_dt.dropna()
    signal_dt = signal_dt.stack()
    signal_dt.name = 'value'
    signal_dt = signal_dt.reset_index()
    ## 信号判断
    ## 0 横盘  1 上涨 2 下跌
    ## 本例子中采用二分法
    #signal_dt['signal'] = np.where(
    #        signal_dt['value'] > 0.3, 1,
    #        np.where(signal_dt['value'] < -0.3, 2, 0))
    signal_dt['signal'] = np.where(
        signal_dt['value'] > 0, 1, 0
    )

    signal_dt['trade_date'] = pd.to_datetime(signal_dt['trade_date'])
    return signal_dt

lable_data = create_lable(copy.deepcopy(market_data))

#### 偏移1天即 T-1的因子簇对应T的涨跌
lable_data['signal'] = lable_data['signal'].shift(-1)
factor_data = factor_data.merge(lable_data,
                                          on=['trade_date', 'code'])

factor_data = factor_data.dropna(subset=['signal'])

#### 保持数据集 用于其他例子

factor_data.to_csv(os.path.join(g_project_data, 'train_datas.csv'),encoding='UTF-8')

##### 重型机器学习模型，目前框架中包含两类 1. XGBTrainer  2. LGBMTrainer 本例子使用 XGBTrainer

model = XGBTrainer(features=features,
                           objective='binary:logistic',
                           booster='gbtree',
                           tree_method='hist',
                           n_estimators=200)

model.fit(factor_data, factor_data['signal'].values)

##### 预测结果

##### 随机构造测试集 用于预测

test_data = pd.DataFrame(np.random.randn(1000, len(features)), columns=features)
test_data.head()

y = model.predict(test_data)
