==> ultron/5.2 time_series_ensemble_learning.ipynb <==
在沙盒环境，将多个机器学习时序模型在集成学习框架下进行涨跌预测

import os,pdb,itertools,copy,datetime
os.environ['ULTRON_DATA'] = 'keim'

import pandas as pd
import numpy as np
from ultron.env import *
from ultron.optimize.model.treemodel import GradientBoostingClassifier
from ultron.optimize.model.treemodel import RandomForestClassifier
from ultron.optimize.model.treemodel import AdaBoostClassifier
from ultron.optimize.model.linearmodel import LogisticRegression

from ultron.optimize.model.treemodel import VotingClassifier
from ultron.optimize.model.treemodel import StackingClassifier

enable_example_env()

#### 加载训练集

train_data = pd.read_csv(os.path.join(g_project_data, 'train_datas.csv'), index_col=0)
train_data.head()

features = [col for col in train_data.columns if col not in ['trade_date','code','inventory','profitratio',
                                                         'value','signal']]

#### 构建训练集

X = train_data[features].fillna(0)
Y = train_data['signal'].values

#### 在实际过程中训练集和测试集要不同数据

#### 训练模型

m1 = RandomForestClassifier(features=features, n_estimators=20)
m1.fit(X, Y)
print(m1.score(X, Y))

m2 = GradientBoostingClassifier(features=features)
m2.fit(X, Y)
print(m2.score(X, Y))

m3 = AdaBoostClassifier(features=features)
m3.fit(X, Y)
print(m3.score(X, Y))

#### 当单体模型训练好后，采用集成学习方式，一共有两种 1.StackingClassifier 2.VotingClassifier

##### VotingClassifier

ensemble1 = VotingClassifier(features=features,
    estimators=[('m1', m1.device), ('m2', m2.device),
                ('m3', m3.device)],voting='soft')
ensemble1.fit(X, Y)
print(ensemble1.score(X, Y))

##### StackingClassifier

ensemble2 = StackingClassifier(
    features=features,
    classifiers=[m1.device, m2.device, m3.device],
    meta_classifier=LogisticRegression().device)
ensemble2.fit(X, Y)
print(ensemble2.score(X, Y))

更多模型训练可参考 optimization_mining_model.ipynb 例子

##### 预测结果

##### 随机构造测试集 用于预测

test_data = pd.DataFrame(np.random.randn(1000, len(features)), columns=features)
test_data.head()

y1 = ensemble1.predict(test_data)

y2 = ensemble2.predict(test_data)
