==> ultron/6.1 optimization_mining_model.ipynb <==
在沙盒环境，通过寻优算法对多种机器学习模型及数据进行寻优，目前选择3种基础机器学习模型 XGBClassifier RandomForestClassifier LGBMClassifier。通过寻优挖掘并选择表现好的前10 precision，放入集成学习模型中

import os,pdb,itertools,copy,datetime
os.environ['ULTRON_DATA'] = 'keim'

import numpy as np
import pandas as pd
from ultron.env import *
from ultron.utilities.utils import NumpyEncoder
from ultron.optimize.geneticist.genetic import Gentic
from ultron.optimize.model.treemodel import XGBClassifier
from ultron.optimize.model.treemodel import LGBMClassifier
from ultron.optimize.model.treemodel import RandomForestClassifier

enable_example_env()

#### 加载训练集

train_data = pd.read_csv(os.path.join(g_project_data, 'train_datas.csv'), index_col=0)
train_data.head()

#### 当前使用3种机器学习模型: XGBClassifier RandomForestClassifier LGBMClassifier

#### 构建模型的参数集

class Parameter(object):
    
    @classmethod
    def XGBClassifier(cls, **kwargs):
        return {
        'max_depth': [i for i in range(6, 15, 2)],
        'n_estimators': [i for i in range(50, 100, 10)],
        'learning_rate': [(i / 100) for i in range(1, 10, 2)],
        'objective': ['multi:softmax'],
        'num_class': [2]}
    
    @classmethod
    def RandomForestClassifier(cls, **kwargs):
        return {
        'n_estimators': [i for i in range(1, 100, 1)],
        'max_features': ['auto']} 
    
    @classmethod
    def LGBMClassifier(cls, **kwargs):
        return {
        'max_depth': [i for i in range(1, 10, 2)],
        'n_estimators': [i for i in range(1, 100, 4)],
        'learning_rate': [(i / 100) for i in range(1, 10, 2)]
    }


params_sets = {
            'XGBClassifier':Parameter.XGBClassifier(),
            'RandomForestClassifier':Parameter.RandomForestClassifier(),
            'LGBMClassifier':Parameter.LGBMClassifier(),
        }

#### 构建基础模型集

model_sets = ['XGBClassifier', 'RandomForestClassifier','LGBMClassifier']

features = [col for col in train_data.columns if col not in ['trade_date','code',
                                                             'value','signal','inventory','profitratio']]

X = train_data[['trade_date','code'] + features].set_index(['trade_date','code']).fillna(0)
Y = train_data[['trade_date','code','signal']].set_index(['trade_date','code']).fillna(0)

#### 创建挖掘引擎

- 1.系统默认3种评估模式，也可自定义datetime估模式
- 2.系统模式评估模式中，已经做了训练集和测试集分离

def save_model(gen, run_details): ## 每一代优秀模型回调
    res = []
    for detail in run_details:
        features = detail._identification
        model_name = detail._model_name
        params = json.dumps(detail._params, cls=NumpyEncoder)
        fitness = detail._raw_fitness
        method = detail._init_method
        res.append({
                'features': features,
                'model_name': model_name,
                'params': params,
                'fitness': fitness,
                'gen': gen,
                'method': method
        }
    )

gentic = Gentic(model_sets=model_sets,
                        params_sets=params_sets,
                        stopping_criteria=100,
                        convergence=0.002,
                        population_size=20,
                        tournament_size=5,
                        p_point_mutation=0.7,
                        p_crossover=0.2,
                        standard_score=0.7,
                        save_model=save_model)

gentic.train(features, X=X, Y=Y, mode='accuracy', n_splits=2) # mode 为模型评估方式
