==> ultron/7.1  machine_learning_training.ipynb <==
在沙盒环境，训练股票的机器学习模型

import os, copy, datetime
os.environ['ULTRON_DATA'] = 'minden'
import pandas as pd
import numpy as np

from ultron.env import *
from ultron.kdutils.file import dump_pickle
from ultron.factor.data.processing import factor_processing
from ultron.factor.data.standardize import standardize
from ultron.factor.data.winsorize import winsorize_normal
from ultron.factor.othgnz.othgnz_engine import OthgnzEngine
from ultron.factor.combine.combine_engine import CombineEngine
from ultron.optimize.model.linearmodel import LinearRegression

from ultron.sentry.Analysis.CrossSectionValueHolders import CSPercentileSecurityValueHolder
from ultron.sentry.Analysis.CrossSectionValueHolders import CSRankedSecurityValueHolder
from ultron.sentry.Analysis.SecurityValueHolders import SecurityDeltaValueHolder
from ultron.sentry.Analysis.SecurityValueHolders import SecurityLatestValueHolder
from ultron.sentry.Analysis.SecurityValueHolders import SecurityShiftedValueHolder
from ultron.sentry.Analysis.TechnicalAnalysis.StatelessTechnicalAnalysers import SecurityLogValueHolder
from ultron.sentry.Analysis.TechnicalAnalysis.StatelessTechnicalAnalysers import SecuritySignValueHolder
from ultron.sentry.Analysis.TechnicalAnalysis.StatefulTechnicalAnalysers import SecurityMovingMax
from ultron.sentry.Analysis.TechnicalAnalysis import SecurityMovingCorrelation
from ultron.sentry.Analysis.TechnicalAnalysis import SecurityMovingRank

enable_example_env()

#### 加载指标数据

indicator_data = pd.read_csv(os.path.join(g_project_data, 'indicator.csv'), index_col=0)
indicator_data['trade_date'] = pd.to_datetime(indicator_data['trade_date'])
indicator_data = indicator_data.rename(columns={'preClosePrice':'pre_close','openPrice':'open',
                        'closePrice':'close','highestPrice':'high','lowestPrice':'low',
                        'turnoverVol':'volume'}).drop(['turnoverValue'],axis=1).set_index(['trade_date'])
indicator_data = indicator_data.sort_values(by=['trade_date','code']).reset_index()
indicator_data.head()

#### 加载行业数据

industry_data = pd.read_csv(os.path.join(g_project_data, 'industry_data.csv'), index_col=0)
industry_data['trade_date'] = pd.to_datetime(industry_data['trade_date'])
industry_data.head()

#### 构建5个alpha191因子作为当前组合因子

def Alpha191_1(data, interval=6):
    exp1 = CSPercentileSecurityValueHolder(
        SecurityDeltaValueHolder(1, SecurityLogValueHolder(SecurityLatestValueHolder('volume'))))
    exp2 = CSPercentileSecurityValueHolder((SecurityLatestValueHolder('close') - \
                SecurityLatestValueHolder('open')) / SecurityLatestValueHolder('open'))

    exp3 = -1 * SecurityMovingCorrelation(interval, exp1, CSRankedSecurityValueHolder(exp2))
    return exp3.transform(data.set_index('trade_date'),name='alpha191_1_' + str(interval) + 'D',
               category_field='code')

def Alpha191_2(data, interval=1):
    exp1 = SecurityLatestValueHolder('close'
                         ) * 2  - SecurityLatestValueHolder('low'
                                                           ) - SecurityLatestValueHolder('high')
    exp2 = SecurityLatestValueHolder('high') - SecurityLatestValueHolder('low')
    exp3 = SecurityDeltaValueHolder(interval, exp1 / exp2) * -1
    return exp3.transform(data.set_index('trade_date'),name='alpha191_2_' + str(interval) + 'D',
               category_field='code')

def Alpha191_5(data, interval=5):
    exp1 = SecurityMovingRank(interval, 'volume')
    exp2 = SecurityMovingRank(interval, 'high')
    exp3 = SecurityMovingCorrelation(interval, exp1 , exp2)
    exp4 = SecurityMovingMax(3, exp3) * -1
    return exp4.transform(data.set_index('trade_date'),name='alpha191_5_' + str(interval) + 'D',
               category_field='code')

def Alpha191_6(data, interval=4):
    exp1 = SecuritySignValueHolder(SecurityDeltaValueHolder(interval,
        SecurityLatestValueHolder('open') * 0.85 + SecurityLatestValueHolder('high') * 0.15))
    
    exp2 = CSPercentileSecurityValueHolder(exp1) * -1
    return exp2.transform(data.set_index('trade_date'),name='alpha191_6_' + str(interval) + 'D',
               category_field='code')


def Alpha191_14(data, interval=1):
    exp1 = SecurityLatestValueHolder('close') - \
SecurityShiftedValueHolder(interval, SecurityLatestValueHolder('close'))
    return exp1.transform(data.set_index('trade_date'),name='alpha191_14_' + str(interval) + 'D',
               category_field='code')

#### 计算因子值

alpha1 = Alpha191_1(indicator_data, interval=6).reset_index().set_index(['trade_date','code'])
alpha2 = Alpha191_2(indicator_data, interval=6).reset_index().set_index(['trade_date','code'])
alpha5 = Alpha191_5(indicator_data, interval=6).reset_index().set_index(['trade_date','code'])
alpha6 = Alpha191_6(indicator_data, interval=6).reset_index().set_index(['trade_date','code'])
alpha14 = Alpha191_14(indicator_data, interval=6).reset_index().set_index(['trade_date','code'])

factors_data = pd.concat([alpha1,alpha2,alpha5,alpha6,alpha14],axis=1).dropna().reset_index()
factors_data.head()

features = [col for col in factors_data.columns if col not in ['trade_date','code']]

diff_cols = ['trade_date','code']
features = [col  for col in factors_data.columns if col not in diff_cols]

neutralized_styles = industry_data['industryID1'].unique().tolist()
industry_dummy = pd.get_dummies(industry_data.set_index(['trade_date','code'])['industryID1']).reset_index()
industry_dummy.head()

factors_data = factors_data.merge(industry_dummy, on=['trade_date','code'])
factors_data.head()

alpha_res = []
grouped = factors_data.groupby(['trade_date'])
for k, g in grouped:
    new_factors = factor_processing(g[features].values,
                 pre_process=[winsorize_normal,standardize],
                 risk_factors=g[neutralized_styles].values.astype(float),
                 post_process=[standardize])
    f = pd.DataFrame(new_factors, columns=features)
    for k in diff_cols:
        f[k] = g[k].values
    alpha_res.append(f)

factors_data = pd.concat(alpha_res)
factors_data.tail()

##### 因子进行合成

symmetry_othgnz = OthgnzEngine.create_engine('symmetry')

factors_data = symmetry_othgnz(
    factors_data[diff_cols + features].fillna(0), 
    diff_cols).sort_values(by=['trade_date','code'])
factors_data.tail()

#### 等权合成

equal_combine = CombineEngine.create_engine('equal_combine')

factors_df = factors_data.copy(deep=True)
factors_df['factor'] = equal_combine(factors_df, features)
equal_data = factors_df.copy(deep=True).drop(features,axis=1).sort_values(by=['trade_date','code'])
equal_data['trade_date'] = pd.to_datetime(equal_data['trade_date'])
equal_data.tail()

#### 构建模型训练

def returns(market_data):
    price_tb = market_data['close'].unstack()
    price_tb.fillna(method='pad', inplace=True)
    return_tb = np.log(price_tb.shift(-1) / price_tb)
    return_tb = return_tb.replace([np.inf, -np.inf], np.nan)
    return_tb = return_tb.stack().reindex(market_data.index)
    return_tb.name = 'nxt1_ret'
    return return_tb

rets = returns(indicator_data.set_index(['trade_date','code']))
rets.head()

train_data = equal_data.merge(rets, on=['trade_date','code'])
train_data = train_data.dropna(subset=['nxt1_ret'])
train_data.head()

model = LinearRegression(['factor'], fit_intercept=False)

model.fit(train_data, train_data['nxt1_ret'].values)

#### 保持模型

dump_pickle(model.save(), os.path.join(g_project_data, 'model.h5'))
