==> ultron/8.3 analysis_of_displacement_distance.ipynb <==
在沙盒环境，分析位移路程分析

import os, copy, sys
os.environ['ULTRON_DATA'] = 'keim'
os.environ['IGNORE_WARNINGS'] = '0'
import pandas as pd
import numpy as np
from matplotlib import pyplot as plt

from ultron.ump.technical.line import Line
from ultron.ump.technical.exectue import EShiftDistanceHow

from ultron.env import *

enable_example_env()

#### 加载行情数据

market_data = pd.read_csv(os.path.join(g_project_data, 'market_data.csv'), index_col=0)
market_data['trade_date'] = pd.to_datetime(market_data['trade_date'])
market_data['date'] = market_data['trade_date']
market_data = market_data.set_index('code').loc['RB'].set_index('trade_date')
market_data = market_data.rename(columns={'closePrice':'close'})[['close','date']]
market_data.tail()

kl_tl = Line(market_data['close'], 'close')

# h_distance, v_distance,         distance,         shift,             sd
# 三角底边距离  三角垂直距离          斜边路程           位移                位移路程比：shift / distance

##### 使用时间序列最后的元素做为路程的计算

shift_distance_list = kl_tl.create_shift_distance(
    how=EShiftDistanceHow.shift_distance_close, step_x=1.0)

##### 使用极限值做为路程的计算基础

shift_distance_list = kl_tl.create_shift_distance(
    how=EShiftDistanceHow.shift_distance_maxmin, step_x=1.0)

##### 使用序列的sum和极限值为路程的计算

shift_distance_list = kl_tl.create_shift_distance(
    how=EShiftDistanceHow.shift_distance_sum_maxmin, step_x=1.0)
