==> ultron/8.10 analysis_of_amplitude.ipynb <==
在沙盒环境下，进行振幅分析

import os, copy, sys
os.environ['ULTRON_DATA'] = 'minden'
os.environ['IGNORE_WARNINGS'] = '0'
import pandas as pd
import numpy as np

from ultron.env import *

from ultron.kdutils.kline import date_week_wave,p_change_stats,wave_change_rate,\
    date_week_win,bcut_change_vc,qcut_change_vc

enable_example_env()

kl_dict = {}
indicator_data = pd.read_csv(os.path.join(g_project_data, 'indicator.csv'), index_col=0)
indicator_data['trade_date'] = pd.to_datetime(indicator_data['trade_date'])
indicator_data = indicator_data.rename(columns={'preClosePrice':'pre_close','openPrice':'open',
                        'closePrice':'close','highestPrice':'high','lowestPrice':'low',
                        'turnoverVol':'volume'}).drop(['turnoverValue'],axis=1).set_index(['trade_date'])
indicator_data['date'] = indicator_data.index
indicator_data['date'] = indicator_data.index
indicator_data['p_change'] = indicator_data['p_change'] * 100
codes = [300002,300012,300017,300558,300618]
for code in codes:
    dt = indicator_data.reset_index().set_index(
    'code').loc[code].reset_index().set_index('trade_date')
    dt.name = str(code)
    kl_dict[str(code)] = dt

##### 交易日振幅对比分析

date_week_wave(kl_dict)

##### 交易日涨跌对比分析

- up_mean:日涨幅平均值
- up_count:上涨走势天数
- down_mean:日跌幅平均值
- up_count:下跌走势天数
- chg_mean:均涨跌比
- chg_ratio:上涨下跌数量比

rval = p_change_stats(kl_dict)
pd.DataFrame(rval)

##### 振幅统计套利条件分析

## 大于1.8 具备套利条件
rval = wave_change_rate(kl_dict)
pd.DataFrame(rval)

##### 交易日涨跌概率分析

date_week_win(kl_dict)

##### 交易日涨跌区间分析(预定区间)

bcut_change_vc(kl_dict)

##### 交易日涨跌区间分析(不定区间)

qcut_change_vc(kl_dict)
