==> ultron/9.2 multi_factor_risk_constraints_of_futures.ipynb <==
在沙盒环境，使用期货多因子组合风险约束

import os,pdb,itertools,copy,datetime,sys,collections
os.environ['ULTRON_DATA'] = 'keim'

import numpy as np
import pandas as pd

from ultron.env import *
from ultron.strategy.builder.factor import ts_zscore_normalization,cross_section_scale
from ultron.strategy.builder.signal import calc_weighted_signal
from ultron.strategy.builder.risk import VolatilityRisk,TurnoverRisk,ExposureRisk,LeverageRisk

enable_example_env()

#### 读取因子

total_data = pd.read_csv(os.path.join(g_project_data, 'factor.csv'), index_col=0)
total_data['trade_date'] = pd.to_datetime(total_data['trade_date'])
factors_data = total_data[['trade_date','code','WeightShortVolRelTotIntChg','BM_RecentSecond_20D']]
factors_data.head()

#### 时序标准化

def normalization(factors_data, name, bound=3, winsize=120):
    factor_data = factors_data.set_index(['trade_date','code'])[name].dropna().unstack()
    rval = ts_zscore_normalization(data=factor_data, bound=bound, winsize=winsize)
    return rval

factor_data1 = normalization(factors_data=factors_data, name='WeightShortVolRelTotIntChg')
factor_data2 = normalization(factors_data=factors_data, name='BM_RecentSecond_20D')

#### 横截面标准化

factor_data1 = cross_section_scale(factor_data1)
factor_data2 = cross_section_scale(factor_data2)

### 多因子信号合成

#### 构建数据集

signals = collections.OrderedDict()

signals['WeightShortVolRelTotIntChg'] = factor_data1
signals['BM_RecentSecond_20D'] = factor_data2
factor_list = ['WeightShortVolRelTotIntChg','BM_RecentSecond_20D']
factor_weight = [0.5,0.5]

#### 因子加权

weigthed = calc_weighted_signal(signals, factor_list, factor_weight)
weigthed.tail()

### 加载行情

market_data = pd.read_csv(os.path.join(g_project_data, 'market_data.csv'), index_col=0)
market_data['trade_date'] = pd.to_datetime(market_data['trade_date'])
market_data.head()

def prev_returs_impl(price_data, key, name):
    price_tb = price_data[key].unstack()
    price_tb.fillna(method='pad', inplace=True)
    return_tb = np.log(price_tb / price_tb.shift(1))
    return_tb = return_tb.replace([np.inf, -np.inf], np.nan)
    return_tb = return_tb.stack().reindex(price_data.index)
    return_tb.name = name
    return return_tb

prev_rets = prev_returs_impl(
    market_data.set_index(['trade_date','code']),'closePrice','pev1_ret').reset_index()
prev_rets['trade_date'] = pd.to_datetime(prev_rets['trade_date'])
prev_rets = prev_rets.set_index(['trade_date', 'code'])['pev1_ret'].unstack()
prev_rets.tail()

#### 波动率约束

volatiliy_risk = VolatilityRisk(volatility_winsize=250, volatility_base=100,
                                upper_limit=1, lower_limit=0.2)
rval = volatiliy_risk.run(portfolio=weigthed,
                          returns=prev_rets,
                          method='portfolio')
rval.tail()

volatiliy_risk = VolatilityRisk(volatility_winsize=250, volatility_base=100,
                                upper_limit=1, lower_limit=0.2)
rval = volatiliy_risk.run(portfolio=weigthed,
                          returns=prev_rets,
                          method='product')
rval.tail()

#### 换手率约束

turnover_risk = TurnoverRisk(grade_step=0.0008)
rval = turnover_risk.run(portfolio=weigthed)
rval.tail()

#### 暴露约束

exposure_risk = ExposureRisk(commission_rate=0.0003,
                             portfolio_exposure=True,
                             max_portfolio=0.1,
                             product_exposure=False,
                             max_product=0.06,
                             turnover_winnratio_winsize=150)
rval = exposure_risk.run(portfolio=weigthed)
rval.tail()

exposure_risk = ExposureRisk(commission_rate=0.0003,
                             portfolio_exposure=False,
                             max_portfolio=0.1,
                             product_exposure=True,
                             max_product=0.06,
                             turnover_winnratio_winsize=150)
rval = exposure_risk.run(portfolio=weigthed)
rval.tail()

#### 杠杆约束

leverage_risk = LeverageRisk(risk_list=[],
                             weight_list=[],
                             risk_builder='riskweighted',
                             risk_control=True)
rval = leverage_risk.run(portfolio=weigthed)
rval.tail()
